+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Updated NFS server template to work with NFS Ganesha instances with monitoring enabled.
+- Ensure Satellites are always rescheduled on NoScheduled nodes as long as the LinstorSatellite resource exists.
 
 ## [v2.10.4] - 2026-01-13